refactor: 레파지토리 패키지를 domain으로 부터 분리
