+++ src/main/java/nextstep/repository/ConsoleReservationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.domain.repository;
+package nextstep.repository;
 
 import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
