+++ src/main/java/nextstep/repository/ReservationRepository.java
@@ -1,4 +1,4 @@
-package nextstep.domain.repository;
+package nextstep.repository;
 
 import nextstep.domain.Reservation;
